Moved eli files into sst/core/eli subdirectory.
