+++ src/sst/core/eli/categoryInfo.h
