+++ src/sst/core/eli/elibase.cc
@@ -11,7 +11,8 @@
 
 
 #include "sst_config.h"
-#include "sst/core/elementinfo.h"
+#include "sst/core/eli/elibase.h"
+#include "sst/core/eli/elementinfo.h"
 #include <sst/core/statapi/statbase.h>
 
 #include <sstream>
