+++ src/sst/core/eli/elibase.h
