+++ src/sst/core/eli/portsInfo.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <string>
 
-#include <sst/core/elibase.h>
+#include <sst/core/eli/elibase.h>
 
 namespace SST {
 namespace ELI {
