+++ src/sst/core/eli/statsInfo.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 #include <string>
-#include <sst/core/elibase.h>
+#include <sst/core/eli/elibase.h>
 
 namespace SST {
 namespace ELI {
