+++ src/sst/core/impl/partitioners/linpart.h
@@ -14,7 +14,7 @@
 #define SST_CORE_IMPL_PARTITONERS_LINPART_H
 
 #include <sst/core/sstpart.h>
-#include <sst/core/elementinfo.h>
+#include <sst/core/eli/elementinfo.h>
 
 namespace SST {
