+++ src/sst/core/impl/partitioners/rrobin.h
@@ -12,7 +12,7 @@
 #define SST_CORE_IMPL_PARTITONERS_RROBIN_H
 
 #include <sst/core/sstpart.h>
-#include <sst/core/elementinfo.h>
+#include <sst/core/eli/elementinfo.h>
 
 namespace SST {
 namespace IMPL {
