+++ src/sst/core/impl/partitioners/singlepart.h
@@ -14,7 +14,7 @@
 #define SST_CORE_IMPL_PARTITONERS_SINGLEPART_H
 
 #include <sst/core/sstpart.h>
-#include <sst/core/elementinfo.h>
+#include <sst/core/eli/elementinfo.h>
 
 namespace SST {
 namespace IMPL {
