+++ src/sst/core/subcomponent.h
@@ -17,6 +17,7 @@
 #include <sst/core/baseComponent.h>
 #include <sst/core/component.h>
 #include <sst/core/module.h>
+#include <sst/core/eli/elementinfo.h>
 
 namespace SST {
